requirements.txt
playwright
requests
python-dotenv
tqdm
